Refactor output files
