+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/output/tripinfo/SumoTripInfoXml.kt
@@ -1,4 +1,4 @@
-package app.urbanflo.urbanflosumoserver.model.output
+package app.urbanflo.urbanflosumoserver.model.output.tripinfo
 
 import app.urbanflo.urbanflosumoserver.simulation.SimulationId
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
