+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/storage/StorageService.kt
@@ -3,9 +3,9 @@ package app.urbanflo.urbanflosumoserver.storage
 import app.urbanflo.urbanflosumoserver.model.SimulationAnalytics
 import app.urbanflo.urbanflosumoserver.model.SimulationInfo
 import app.urbanflo.urbanflosumoserver.model.network.SumoNetwork
-import app.urbanflo.urbanflosumoserver.model.output.SumoNetstateXml
+import app.urbanflo.urbanflosumoserver.model.output.netstate.SumoNetstateXml
 import app.urbanflo.urbanflosumoserver.model.output.SumoSimulationOutput
-import app.urbanflo.urbanflosumoserver.model.output.SumoTripInfoXml
+import app.urbanflo.urbanflosumoserver.model.output.tripinfo.SumoTripInfoXml
 import app.urbanflo.urbanflosumoserver.simulation.SimulationId
 import app.urbanflo.urbanflosumoserver.simulation.SimulationInstance
